reorganizing things
