--- S2CS/README.md
@@ -1,8 +1,0 @@
-# SciStream Control Server (S2CS)
-
-The main functions of S2CS are to authenticate users and ProdAPP/ConsAPP, to interact (via authenticated control channels) with S2UC and ProdAPP/ConsAPP in order to establish authenticated data channels, and to interact with S2DS initiator and S2DS in order to create the connection map between producer and consumer gateway nodes and manage the connections.
-S2CS is a persistently running process.
-
-## State Machine
-
-![alt text](../figures/s2cs-state-machine.png "S2CS state machine")
