+++ pyproject.toml
@@ -8,8 +8,6 @@ packages = [{include = "scistream_proto"}]
 
 [tool.poetry.dependencies]
 python = "^3.7"
-zmq = "^0.0.0"
-transitions = "^0.9.0"
 optparse-pretty = "^0.1.1"
 grpcio = "^1.51.3"
 grpcio-tools = "^1.51.3"
@@ -18,6 +16,7 @@ click = "^8.1.3"
 fire = "^0.5.0"
 grpcio-testing = "^1.53.0"
 pytest-timeout = "^2.1.0"
+protobuf = "^4.22.3"
 
 
 [tool.poetry.group.test.dependencies]
