+++ src/proto/scistream_pb2_grpc.py
@@ -2,7 +2,7 @@
 """Client and server classes corresponding to protobuf-defined services."""
 import grpc
 
-import scistream_pb2 as scistream__pb2
+import proto.scistream_pb2 as scistream__pb2
 
 
 class ControlStub(object):
