--- utils/README.md
@@ -1,15 +1,0 @@
-# Utils
-
-This folder contains utilities and state machines for entities not part of SciStream protocol (i.e., user, ProdApp, and ConsApp), but still necessary for its well functioning.
-
-## User State Machine
-
-![alt text](../figures/user-state-machine.png "User state machine")
-
-## ProdApp State Machine
-
-![alt text](../figures/prodapp-state-machine.png "ProdApp state machine")
-
-## ConsApp State Machine
-
-![alt text](../figures/consapp-state-machine.png "ConsApp state machine")
